Changes to the Events System
The events system in place was only groundwork. It was functional but because of its design it was not going to be usable later on in the scripting system. Now more finite control over events is possible, and a mechanism for attaching events to specific users was added (although a system for registering user events has not been implemented yet).

I have documented the Events system in the github wiki for the project, but further changes are likely coming. Work will continue on this system until it is stable enough to support the forthcoming scripting system.